+++ src/animation/Display.ts
@@ -1,5 +1,5 @@
 import { AnimationDisplayConfig } from '@/types/Config';
-import { ImageObject } from '../types/Gullview';
+import { ImageObject } from '@/types/Gullview';
 
 import defaults from '@/config/defaults';
